requirements.txt
requests>=2.31.0
aiohttp>=3.8.0
asyncio>=3.4.3
click>=8.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0